restructure server
