+++ packages/server/middleware/routes/index.js
@@ -1,4 +1,3 @@
-const screenRoutes = require("./screens");
 const pageRoutes = require("./pages");
 const componentRoutes = require("./components");
 const userRoutes = require("./user");
@@ -6,9 +5,9 @@ const appsRoutes = require("./apps");
 const authenticatedRoutes = require("./authenticated");
 
 module.exports = {
-  screenRoutes,
   pageRoutes,
   componentRoutes,
-  authRoutes,
-  appsRoutes
+  appsRoutes,
+  userRoutes,
+  authenticatedRoutes
 };
